requirements.txt
azure-core==1.10.0
azure-functions==1.6.0
azure-storage-blob==12.7.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
idna==2.10
isodate==0.6.0
joblib==1.0.1
lxml==4.6.2
msrest==0.6.21
numpy==1.20.0
oauthlib==3.1.0
pandas==1.2.1
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
soupsieve==2.1
threadpoolctl==2.1.0
urllib3==1.26.3
